librz/core/*.c refactor for RzTypes extraction
