+++ librz/core/cmd_type.c
@@ -7,6 +7,264 @@
 #include <sdb.h>
 #include "core_private.h"
 
+// Calling conventions
+
+static void types_cc_print(RzCore *core, const char *cc, RzOutputMode mode) {
+	rz_return_if_fail(cc);
+	if (strchr(cc, '(')) {
+		if (!rz_analysis_cc_set(core->analysis, cc)) {
+			eprintf("Invalid syntax in cc signature.");
+		}
+	} else {
+		const char *ccname = rz_str_trim_head_ro(cc);
+		char *result = rz_analysis_cc_get(core->analysis, ccname);
+		if (result) {
+			if (mode == RZ_OUTPUT_MODE_JSON) {
+				PJ *pj = rz_core_pj_new(core);
+				pj_a(pj);
+				pj_ks(pj, "cc", result);
+				pj_end(pj);
+				rz_cons_println(pj_string(pj));
+				pj_free(pj);
+			} else {
+				rz_cons_printf("%s\n", result);
+			}
+			free(result);
+		}
+	}
+}
+
+// Enums
+
+static RzCmdStatus types_enum_member_find(RzCore *core, const char *enum_name, const char *enum_value) {
+	rz_return_val_if_fail(enum_name || enum_value, RZ_CMD_STATUS_ERROR);
+	Sdb *TDB = core->analysis->sdb_types;
+	ut64 value = rz_num_math(core->num, enum_value);
+	char *enum_member = rz_type_enum_member(TDB, enum_name, NULL, value);
+	if (!enum_member) {
+		eprintf("Cannot find matching enum member");
+		return RZ_CMD_STATUS_ERROR;
+	}
+	rz_cons_println(enum_member);
+	free(enum_member);
+	return RZ_CMD_STATUS_OK;
+}
+
+static RzCmdStatus types_enum_member_find_all(RzCore *core, const char *enum_value) {
+	rz_return_val_if_fail(enum_value, RZ_CMD_STATUS_ERROR);
+	Sdb *TDB = core->analysis->sdb_types;
+	ut64 value = rz_num_math(core->num, enum_value);
+	RzList *matches = rz_type_enum_find_member(TDB, value);
+	if (!matches || rz_list_empty(matches)) {
+		eprintf("Cannot find matching enum member");
+		return RZ_CMD_STATUS_ERROR;
+	}
+	RzListIter *iter;
+	char *match;
+	rz_list_foreach (matches, iter, match) {
+		rz_cons_println(match);
+	}
+	rz_list_free(matches);
+	return RZ_CMD_STATUS_OK;
+}
+
+static bool print_type_c(RzCore *core, const char *ctype) {
+	Sdb *TDB = core->analysis->sdb_types;
+	const char *type = rz_str_trim_head_ro(ctype);
+	const char *name = type ? strchr(type, '.') : NULL;
+	if (name && type) {
+		name++; // skip the '.'
+		if (rz_str_startswith(type, "struct")) {
+			rz_types_struct_print_c(TDB, name, true);
+		} else if (rz_str_startswith(type, "union")) {
+			rz_types_union_print_c(TDB, name, true);
+		} else if (rz_str_startswith(type, "enum")) {
+			rz_types_enum_print_c(TDB, name, true);
+		} else if (rz_str_startswith(type, "typedef")) {
+			rz_types_typedef_print_c(TDB, name);
+		} else if (rz_str_startswith(type, "func")) {
+			rz_types_function_print(TDB, name, RZ_OUTPUT_MODE_STANDARD, NULL);
+		}
+		return true;
+	}
+	return false;
+}
+
+static void type_list_c_all(RzCore *core) {
+	Sdb *TDB = core->analysis->sdb_types;
+	// List all unions in the C format with newlines
+	rz_types_union_print_c(TDB, NULL, true);
+	// List all structures in the C format with newlines
+	rz_types_struct_print_c(TDB, NULL, true);
+	// List all typedefs in the C format with newlines
+	rz_types_typedef_print_c(TDB, NULL);
+	// List all enums in the C format with newlines
+	rz_types_enum_print_c(TDB, NULL, true);
+}
+
+static void type_list_c_all_nl(RzCore *core) {
+	Sdb *TDB = core->analysis->sdb_types;
+	// List all unions in the C format without newlines
+	rz_types_union_print_c(TDB, NULL, false);
+	// List all structures in the C format without newlines
+	rz_types_struct_print_c(TDB, NULL, false);
+	// List all typedefs in the C format without newlines
+	rz_types_typedef_print_c(TDB, NULL);
+	// List all enums in the C format without newlines
+	rz_types_enum_print_c(TDB, NULL, false);
+}
+
+static RzCmdStatus type_format_print(RzCore *core, const char *type, ut64 address) {
+	Sdb *TDB = core->analysis->sdb_types;
+	char *fmt = rz_type_format(TDB, type);
+	if (RZ_STR_ISEMPTY(fmt)) {
+		eprintf("Cannot find type %s\n", type);
+		return RZ_CMD_STATUS_ERROR;
+	}
+	rz_core_cmdf(core, "pf %s @ 0x%08" PFMT64x "\n", fmt, address);
+	return RZ_CMD_STATUS_OK;
+}
+
+static RzCmdStatus type_format_print_variable(RzCore *core, const char *type, const char *varname) {
+	Sdb *TDB = core->analysis->sdb_types;
+	char *fmt = rz_type_format(TDB, type);
+	if (RZ_STR_ISEMPTY(fmt)) {
+		eprintf("Cannot find type \"%s\"\n", type);
+		return RZ_CMD_STATUS_ERROR;
+	}
+	RzAnalysisFunction *fcn = rz_analysis_get_fcn_in(core->analysis, core->offset, -1);
+	if (!fcn) {
+		eprintf("Cannot find function at the current offset\n");
+		return RZ_CMD_STATUS_ERROR;
+	}
+	RzAnalysisVar *var = rz_analysis_function_get_var_byname(fcn, varname);
+	if (!var) {
+		eprintf("Cannot find variable \"%s\" in the current function\n", varname);
+		return RZ_CMD_STATUS_ERROR;
+	}
+	ut64 addr = rz_analysis_var_addr(var);
+	rz_core_cmdf(core, "pf %s @ 0x%08" PFMT64x "\n", fmt, addr);
+	return RZ_CMD_STATUS_OK;
+}
+
+static RzCmdStatus type_format_print_value(RzCore *core, const char *type, ut64 val) {
+	Sdb *TDB = core->analysis->sdb_types;
+	char *fmt = rz_type_format(TDB, type);
+	if (RZ_STR_ISEMPTY(fmt)) {
+		eprintf("Cannot find type %s\n", type);
+		return RZ_CMD_STATUS_ERROR;
+	}
+	rz_core_cmdf(core, "pf %s @v:0x%08" PFMT64x "\n", fmt, val);
+	return RZ_CMD_STATUS_OK;
+}
+
+static RzCmdStatus type_format_print_hexstring(RzCore *core, const char *type, const char *hexpairs) {
+	Sdb *TDB = core->analysis->sdb_types;
+	char *fmt = rz_type_format(TDB, type);
+	if (RZ_STR_ISEMPTY(fmt)) {
+		eprintf("Cannot find type %s\n", type);
+		return RZ_CMD_STATUS_ERROR;
+	}
+	rz_core_cmdf(core, "pf %s @x:%s", fmt, hexpairs);
+	return RZ_CMD_STATUS_OK;
+}
+
+static void type_define(RzCore *core, const char *type) {
+	// Add trailing semicolon to force the valid C syntax
+	// It allows us to skip the trailing semicolon in the input
+	// to reduce the unnecessary typing
+	char *tmp = rz_str_newf("%s;", type);
+	if (!tmp) {
+		return;
+	}
+	char *error_msg = NULL;
+	char *out = rz_parse_c_string(core->analysis, tmp, &error_msg);
+	free(tmp);
+	if (out) {
+		rz_analysis_save_parsed_type(core->analysis, out);
+		free(out);
+	}
+	if (error_msg) {
+		eprintf("%s", error_msg);
+		free(error_msg);
+	}
+}
+
+static void types_open_file(RzCore *core, const char *path) {
+	const char *dir = rz_config_get(core->config, "dir.types");
+	char *homefile = NULL;
+	if (*path == '~') {
+		if (path[1] && path[2]) {
+			homefile = rz_str_home(path + 2);
+			path = homefile;
+		}
+	}
+	if (!strcmp(path, "-")) {
+		char *tmp = rz_core_editor(core, "*.h", "");
+		if (tmp) {
+			char *error_msg = NULL;
+			char *out = rz_parse_c_string(core->analysis, tmp, &error_msg);
+			if (out) {
+				rz_analysis_save_parsed_type(core->analysis, out);
+				free(out);
+			}
+			if (error_msg) {
+				fprintf(stderr, "%s", error_msg);
+				free(error_msg);
+			}
+			free(tmp);
+		}
+	} else {
+		char *error_msg = NULL;
+		char *out = rz_parse_c_file(core->analysis, path, dir, &error_msg);
+		if (out) {
+			rz_analysis_save_parsed_type(core->analysis, out);
+			free(out);
+		}
+		if (error_msg) {
+			fprintf(stderr, "%s", error_msg);
+			free(error_msg);
+		}
+	}
+	free(homefile);
+}
+
+static void types_open_editor(RzCore *core, const char *typename) {
+	Sdb *TDB = core->analysis->sdb_types;
+	char *str = rz_core_cmd_strf(core, "tc %s", typename ? typename : "");
+	char *tmp = rz_core_editor(core, "*.h", str);
+	if (tmp) {
+		char *error_msg = NULL;
+		char *out = rz_parse_c_string(core->analysis, tmp, &error_msg);
+		if (out) {
+			// remove previous types and save new edited types
+			sdb_reset(TDB);
+			rz_parse_c_reset(core->parser);
+			rz_analysis_save_parsed_type(core->analysis, out);
+			free(out);
+		}
+		if (error_msg) {
+			eprintf("%s\n", error_msg);
+			free(error_msg);
+		}
+		free(tmp);
+	}
+	free(str);
+}
+
+static void types_open_sdb(RzCore *core, const char *path) {
+	Sdb *TDB = core->analysis->sdb_types;
+	if (rz_file_exists(path)) {
+		Sdb *db_tmp = sdb_new(0, path, 0);
+		sdb_merge(TDB, db_tmp);
+		sdb_close(db_tmp);
+		sdb_free(db_tmp);
+	}
+}
+
+// =============================================================================
+//                             DEPRECATED
+
 static const char *help_msg_t[] = {
 	"Usage: t", "", "# cparse types commands",
 	"t", "", "List all loaded types",
@@ -175,368 +433,6 @@ static const char *help_msg_tu[] = {
 	NULL
 };
 
-static void kv_lines_print_sorted(char *kv_lines) {
-	RzListIter *iter;
-	char *k;
-	RzList *list = rz_str_split_duplist(kv_lines, "\n", true);
-	rz_list_sort(list, (RzListComparator)strcmp);
-	rz_list_foreach (list, iter, k) {
-		if (RZ_STR_ISNOTEMPTY(k)) {
-			rz_cons_println(k);
-		}
-	}
-	rz_list_free(list);
-}
-
-static void types_cc_print_all(RzCore *core, RzOutputMode mode) {
-	switch (mode) {
-	case RZ_OUTPUT_MODE_STANDARD: {
-		rz_core_analysis_calling_conventions_print(core);
-	} break;
-	case RZ_OUTPUT_MODE_JSON: {
-		RzList *list = rz_core_analysis_calling_conventions(core);
-		RzListIter *iter;
-		const char *cc;
-		PJ *pj = rz_core_pj_new(core);
-		pj_a(pj);
-		rz_list_foreach (list, iter, cc) {
-			char *ccexpr = rz_analysis_cc_get(core->analysis, cc);
-			// TODO: expose this as an object, not just an array of strings
-			pj_s(pj, ccexpr);
-			free(ccexpr);
-		}
-		pj_end(pj);
-		rz_cons_printf("%s\n", pj_string(pj));
-		pj_free(pj);
-		rz_list_free(list);
-	} break;
-	case RZ_OUTPUT_MODE_LONG: {
-		RzList *list = rz_core_analysis_calling_conventions(core);
-		RzListIter *iter;
-		const char *cc;
-		rz_list_foreach (list, iter, cc) {
-			char *ccexpr = rz_analysis_cc_get(core->analysis, cc);
-			rz_cons_printf("%s\n", ccexpr);
-			free(ccexpr);
-		}
-		rz_list_free(list);
-	} break;
-	case RZ_OUTPUT_MODE_RIZIN: {
-		RzList *list = rz_core_analysis_calling_conventions(core);
-		RzListIter *iter;
-		const char *cc;
-		rz_list_foreach (list, iter, cc) {
-			char *ccexpr = rz_analysis_cc_get(core->analysis, cc);
-			rz_cons_printf("tcc \"%s\"\n", ccexpr);
-			free(ccexpr);
-		}
-		rz_list_free(list);
-	} break;
-	case RZ_OUTPUT_MODE_SDB:
-		rz_core_kuery_print(core, "analysis/cc/*");
-		break;
-	default:
-		rz_warn_if_reached();
-		break;
-	}
-}
-
-static void types_cc_print(RzCore *core, const char *cc, RzOutputMode mode) {
-	rz_return_if_fail(cc);
-	if (strchr(cc, '(')) {
-		if (!rz_analysis_cc_set(core->analysis, cc)) {
-			eprintf("Invalid syntax in cc signature.");
-		}
-	} else {
-		const char *ccname = rz_str_trim_head_ro(cc);
-		char *result = rz_analysis_cc_get(core->analysis, ccname);
-		if (result) {
-			if (mode == RZ_OUTPUT_MODE_JSON) {
-				PJ *pj = rz_core_pj_new(core);
-				pj_a(pj);
-				pj_ks(pj, "cc", result);
-				pj_end(pj);
-				rz_cons_println(pj_string(pj));
-				pj_free(pj);
-			} else {
-				rz_cons_printf("%s\n", result);
-			}
-			free(result);
-		}
-	}
-}
-
-static void types_enum_print(RzCore *core, const char *enum_name, RzOutputMode mode, PJ *pj) {
-	rz_return_if_fail(enum_name);
-	Sdb *TDB = core->analysis->sdb_types;
-	switch (mode) {
-	case RZ_OUTPUT_MODE_JSON: {
-		rz_return_if_fail(pj);
-		RTypeEnum *member;
-		RzListIter *iter;
-		RzList *list = rz_type_get_enum(TDB, enum_name);
-		pj_o(pj);
-		if (list && !rz_list_empty(list)) {
-			pj_ks(pj, "name", enum_name);
-			pj_k(pj, "values");
-			pj_o(pj);
-			rz_list_foreach (list, iter, member) {
-				pj_kn(pj, member->name, rz_num_math(NULL, member->val));
-			}
-			pj_end(pj);
-		}
-		pj_end(pj);
-		rz_list_free(list);
-		break;
-	}
-	case RZ_OUTPUT_MODE_STANDARD: {
-		RzList *list = rz_type_get_enum(TDB, enum_name);
-		RzListIter *iter;
-		RTypeEnum *member;
-		rz_list_foreach (list, iter, member) {
-			rz_cons_printf("%s = %s\n", member->name, member->val);
-		}
-		rz_list_free(list);
-		break;
-	}
-	case RZ_OUTPUT_MODE_QUIET:
-		rz_cons_println(enum_name);
-		break;
-	case RZ_OUTPUT_MODE_SDB: {
-		char *keys = sdb_querys(TDB, NULL, -1, sdb_fmt("~~enum.%s", enum_name));
-		if (keys) {
-			kv_lines_print_sorted(keys);
-			free(keys);
-		}
-		break;
-	}
-	default:
-		break;
-	}
-}
-
-static void types_enum_print_all(RzCore *core, RzOutputMode mode) {
-	Sdb *TDB = core->analysis->sdb_types;
-	SdbKv *kv;
-	SdbListIter *iter;
-	PJ *pj = (mode == RZ_OUTPUT_MODE_JSON) ? rz_core_pj_new(core) : NULL;
-	SdbList *l = sdb_foreach_list(TDB, true);
-	if (mode == RZ_OUTPUT_MODE_JSON) {
-		pj_a(pj);
-	}
-	ls_foreach (l, iter, kv) {
-		if (!strcmp(sdbkv_value(kv), "enum")) {
-			const char *name = sdbkv_key(kv);
-			types_enum_print(core, name, mode, pj);
-		}
-	}
-	ls_free(l);
-	if (mode == RZ_OUTPUT_MODE_JSON) {
-		pj_end(pj);
-		rz_cons_println(pj_string(pj));
-		pj_free(pj);
-	}
-}
-
-static RzCmdStatus types_enum_member_find(RzCore *core, const char *enum_name, const char *enum_value) {
-	rz_return_val_if_fail(enum_name || enum_value, RZ_CMD_STATUS_ERROR);
-	Sdb *TDB = core->analysis->sdb_types;
-	ut64 value = rz_num_math(core->num, enum_value);
-	char *enum_member = rz_type_enum_member(TDB, enum_name, NULL, value);
-	if (!enum_member) {
-		eprintf("Cannot find matching enum member");
-		return RZ_CMD_STATUS_ERROR;
-	}
-	rz_cons_println(enum_member);
-	free(enum_member);
-	return RZ_CMD_STATUS_OK;
-}
-
-static RzCmdStatus types_enum_member_find_all(RzCore *core, const char *enum_value) {
-	rz_return_val_if_fail(enum_value, RZ_CMD_STATUS_ERROR);
-	Sdb *TDB = core->analysis->sdb_types;
-	ut64 value = rz_num_math(core->num, enum_value);
-	RzList *matches = rz_type_enum_find_member(TDB, value);
-	if (!matches || rz_list_empty(matches)) {
-		eprintf("Cannot find matching enum member");
-		return RZ_CMD_STATUS_ERROR;
-	}
-	RzListIter *iter;
-	char *match;
-	rz_list_foreach (matches, iter, match) {
-		rz_cons_println(match);
-	}
-	rz_list_free(matches);
-	return RZ_CMD_STATUS_OK;
-}
-
-static void types_function_print(RzCore *core, const char *function, RzOutputMode mode, PJ *pj) {
-	rz_return_if_fail(function);
-	Sdb *TDB = core->analysis->sdb_types;
-	char *res = sdb_querys(TDB, NULL, -1, sdb_fmt("func.%s.args", function));
-	int i, args = sdb_num_get(TDB, sdb_fmt("func.%s.args", function), 0);
-	const char *ret = sdb_const_get(TDB, sdb_fmt("func.%s.ret", function), 0);
-	if (!ret) {
-		ret = "void";
-	}
-	switch (mode) {
-	case RZ_OUTPUT_MODE_JSON: {
-		rz_return_if_fail(pj);
-		pj_o(pj);
-		pj_ks(pj, "name", function);
-		pj_ks(pj, "ret", ret);
-		pj_k(pj, "args");
-		pj_a(pj);
-		for (i = 0; i < args; i++) {
-			char *type = sdb_get(TDB, sdb_fmt("func.%s.arg.%d", function, i), 0);
-			if (!type) {
-				continue;
-			}
-			char *name = strchr(type, ',');
-			if (name) {
-				*name++ = 0;
-			}
-			pj_o(pj);
-			pj_ks(pj, "type", type);
-			if (name) {
-				pj_ks(pj, "name", name);
-			} else {
-				pj_ks(pj, "name", "(null)");
-			}
-			pj_end(pj);
-		}
-		pj_end(pj);
-		pj_end(pj);
-	} break;
-	case RZ_OUTPUT_MODE_SDB: {
-		char *keys = sdb_querys(TDB, NULL, -1, sdb_fmt("~~func.%s", function));
-		if (keys) {
-			kv_lines_print_sorted(keys);
-			free(keys);
-		}
-	} break;
-	default: {
-		rz_cons_printf("%s %s (", ret, function);
-		for (i = 0; i < args; i++) {
-			char *type = sdb_get(TDB, sdb_fmt("func.%s.arg.%d", function, i), 0);
-			char *name = strchr(type, ',');
-			if (name) {
-				*name++ = 0;
-			}
-			rz_cons_printf("%s%s %s", i == 0 ? "" : ", ", type, name);
-		}
-		rz_cons_printf(");\n");
-	} break;
-	}
-	free(res);
-}
-
-static void types_function_print_all(RzCore *core, RzOutputMode mode) {
-	Sdb *TDB = core->analysis->sdb_types;
-	SdbKv *kv;
-	SdbListIter *iter;
-	PJ *pj = (mode == RZ_OUTPUT_MODE_JSON) ? rz_core_pj_new(core) : NULL;
-	SdbList *l = sdb_foreach_list(TDB, true);
-	if (mode == RZ_OUTPUT_MODE_JSON) {
-		pj_a(pj);
-	}
-	ls_foreach (l, iter, kv) {
-		if (!strcmp(sdbkv_value(kv), "func")) {
-			const char *name = sdbkv_key(kv);
-			types_function_print(core, name, mode, pj);
-		}
-	}
-	ls_free(l);
-	if (mode == RZ_OUTPUT_MODE_JSON) {
-		pj_end(pj);
-		rz_cons_println(pj_string(pj));
-		pj_free(pj);
-	}
-}
-
-static void types_link_print(RzCore *core, const char *type, ut64 addr, RzOutputMode mode) {
-	rz_return_if_fail(type);
-	switch (mode) {
-	case RZ_OUTPUT_MODE_JSON: {
-		PJ *pj = pj_new();
-		if (!pj) {
-			return;
-		}
-		pj_o(pj);
-		char *saddr = rz_str_newf("0x%" PFMT64x, addr);
-		pj_ks(pj, saddr, type);
-		pj_end(pj);
-		rz_cons_println(pj_string(pj));
-		pj_free(pj);
-		free(saddr);
-		break;
-	}
-	case RZ_OUTPUT_MODE_STANDARD:
-		rz_cons_printf("0x%" PFMT64x " = %s\n", addr, type);
-		break;
-	case RZ_OUTPUT_MODE_RIZIN:
-		rz_cons_printf("tl %s 0x%" PFMT64x "\n", type, addr);
-		break;
-	case RZ_OUTPUT_MODE_LONG: {
-		char *fmt = rz_type_format(core->analysis->sdb_types, type);
-		if (!fmt) {
-			eprintf("Can't fint type %s", type);
-		}
-		rz_cons_printf("(%s)\n", type);
-		rz_core_cmdf(core, "pf %s @ 0x%" PFMT64x "\n", fmt, addr);
-		break;
-	}
-	default:
-		rz_warn_if_reached();
-		break;
-	}
-}
-
-static void types_link_print_all(RzCore *core, RzOutputMode mode) {
-	Sdb *TDB = core->analysis->sdb_types;
-	SdbKv *kv;
-	SdbListIter *iter;
-	SdbList *l = sdb_foreach_list(TDB, true);
-	ls_foreach (l, iter, kv) {
-		if (!strcmp(sdbkv_value(kv), "link")) {
-			const char *name = sdbkv_key(kv);
-			ut64 addr = rz_num_math(core->num, sdbkv_value(kv) + strlen("link."));
-			types_link_print(core, name, addr, mode);
-		}
-	}
-	ls_free(l);
-}
-
-static void types_link(RzCore *core, const char *type, ut64 addr) {
-	Sdb *TDB = core->analysis->sdb_types;
-	char *tmp = sdb_get(TDB, type, 0);
-	if (tmp && *tmp) {
-		rz_type_set_link(TDB, type, addr);
-		RzList *fcns = rz_analysis_get_functions_in(core->analysis, core->offset);
-		if (rz_list_length(fcns) > 1) {
-			eprintf("Multiple functions found in here.\n");
-		} else if (rz_list_length(fcns) == 1) {
-			RzAnalysisFunction *fcn = rz_list_first(fcns);
-			rz_core_link_stroff(core, fcn);
-		} else {
-			eprintf("Cannot find any function here\n");
-		}
-		rz_list_free(fcns);
-		free(tmp);
-	} else {
-		eprintf("unknown type %s\n", type);
-	}
-}
-
-static void types_link_show(RzCore *core, ut64 addr) {
-	Sdb *TDB = core->analysis->sdb_types;
-	const char *query = sdb_fmt("link.%08" PFMT64x, addr);
-	const char *link = sdb_const_get(TDB, query, 0);
-	if (link) {
-		types_link_print(core, link, addr, RZ_OUTPUT_MODE_LONG);
-	}
-}
-
 static void __core_cmd_tcc(RzCore *core, const char *input) {
 	switch (*input) {
 	case '?':
@@ -550,62 +446,23 @@ static void __core_cmd_tcc(RzCore *core, const char *input) {
 		}
 		break;
 	case 0:
-		types_cc_print_all(core, RZ_OUTPUT_MODE_STANDARD);
+		rz_core_types_calling_conventions_print(core, RZ_OUTPUT_MODE_STANDARD);
 		break;
 	case 'j':
-		types_cc_print_all(core, RZ_OUTPUT_MODE_JSON);
+		rz_core_types_calling_conventions_print(core, RZ_OUTPUT_MODE_JSON);
 		break;
 	case 'l':
-		types_cc_print_all(core, RZ_OUTPUT_MODE_LONG);
+		rz_core_types_calling_conventions_print(core, RZ_OUTPUT_MODE_LONG);
 		break;
 	case '*':
-		types_cc_print_all(core, RZ_OUTPUT_MODE_RIZIN);
+		rz_core_types_calling_conventions_print(core, RZ_OUTPUT_MODE_RIZIN);
 		break;
 	case 'k':
-		types_cc_print_all(core, RZ_OUTPUT_MODE_SDB);
+		rz_core_types_calling_conventions_print(core, RZ_OUTPUT_MODE_SDB);
 		break;
 	case ' ':
 		types_cc_print(core, input + 1, RZ_OUTPUT_MODE_STANDARD);
 		break;
-	}
-}
-
-static void type_show_format(RzCore *core, const char *name, RzOutputMode mode) {
-	const char *isenum = sdb_const_get(core->analysis->sdb_types, name, 0);
-	if (isenum && !strcmp(isenum, "enum")) {
-		eprintf("IS ENUM\n");
-	} else {
-		char *fmt = rz_type_format(core->analysis->sdb_types, name);
-		if (fmt) {
-			rz_str_trim(fmt);
-			switch (mode) {
-			case RZ_OUTPUT_MODE_JSON: {
-				PJ *pj = rz_core_pj_new(core);
-				if (!pj) {
-					return;
-				}
-				pj_o(pj);
-				pj_ks(pj, "name", name);
-				pj_ks(pj, "format", fmt);
-				pj_end(pj);
-				rz_cons_printf("%s", pj_string(pj));
-				pj_free(pj);
-			} break;
-			case RZ_OUTPUT_MODE_RIZIN: {
-				rz_cons_printf("pf.%s %s\n", name, fmt);
-			} break;
-			case RZ_OUTPUT_MODE_STANDARD: {
-				// FIXME: Not really a standard format
-				// We should think about better representation by default here
-				rz_cons_printf("pf %s\n", fmt);
-			} break;
-			default:
-				break;
-			}
-			free(fmt);
-		} else {
-			eprintf("Cannot find '%s' type\n", name);
-		}
 	}
 }
 
@@ -623,7 +480,7 @@ static void cmd_type_noreturn(RzCore *core, const char *input) {
 	switch (input[0]) {
 	case '-': // "tn-"
 		if (input[1] == '*') {
-			RzList *noretl = rz_core_analysis_noreturn(core);
+			RzList *noretl = rz_types_function_noreturn(core->analysis->sdb_types);
 			RzListIter *iter;
 			char *name;
 			rz_list_foreach (noretl, iter, name) {
@@ -663,13 +520,13 @@ static void cmd_type_noreturn(RzCore *core, const char *input) {
 		break;
 	case '*':
 	case 'r': // "tn*"
-		rz_core_analysis_noreturn_print(core, RZ_OUTPUT_MODE_RIZIN);
+		rz_core_types_function_noreturn_print(core, RZ_OUTPUT_MODE_RIZIN);
 		break;
 	case 'j': // "tnj"
-		rz_core_analysis_noreturn_print(core, RZ_OUTPUT_MODE_JSON);
+		rz_core_types_function_noreturn_print(core, RZ_OUTPUT_MODE_JSON);
 		break;
 	case 0: // "tn"
-		rz_core_analysis_noreturn_print(core, RZ_OUTPUT_MODE_STANDARD);
+		rz_core_types_function_noreturn_print(core, RZ_OUTPUT_MODE_STANDARD);
 		break;
 	default:
 	case '?':
@@ -678,766 +535,18 @@ static void cmd_type_noreturn(RzCore *core, const char *input) {
 	}
 }
 
-static Sdb *TDB_ = NULL; // HACK
-
-static bool stdifstruct(void *user, const char *k, const char *v) {
-	rz_return_val_if_fail(TDB_, false);
-	if (!strcmp(v, "struct") && !rz_str_startswith(k, "typedef")) {
-		return true;
-	}
-	if (!strcmp(v, "typedef")) {
-		const char *typedef_key = sdb_fmt("typedef.%s", k);
-		const char *type = sdb_const_get(TDB_, typedef_key, NULL);
-		if (type && rz_str_startswith(type, "struct")) {
-			return true;
-		}
-	}
-	return false;
-}
-
-/*!
- * \brief print the data types details in JSON format
- * \param TDB pointer to the sdb for types
- * \param filter a callback function for the filtering
- * \return 1 if success, 0 if failure
- */
-static int print_struct_union_list_json(Sdb *TDB, SdbForeachCallback filter) {
-	PJ *pj = pj_new();
-	if (!pj) {
-		return 0;
-	}
-	SdbList *l = sdb_foreach_list_filter(TDB, filter, true);
-	SdbListIter *it;
-	SdbKv *kv;
-
-	pj_a(pj); // [
-	ls_foreach (l, it, kv) {
-		const char *k = sdbkv_key(kv);
-		if (!k || !*k) {
-			continue;
-		}
-		pj_o(pj); // {
-		pj_ks(pj, "type", k); // key value pair of string and string
-		pj_end(pj); // }
-	}
-	pj_end(pj); // ]
-
-	rz_cons_println(pj_string(pj));
-	pj_free(pj);
-	ls_free(l);
-	return 1;
-}
-
-static void print_struct_union_in_c_format(Sdb *TDB, SdbForeachCallback filter, const char *arg, bool multiline) {
-	char *name = NULL;
-	SdbKv *kv;
-	SdbListIter *iter;
-	SdbList *l = sdb_foreach_list_filter(TDB, filter, true);
-	const char *space = "";
-	bool match = false;
-
-	ls_foreach (l, iter, kv) {
-		if (name && !strcmp(sdbkv_value(kv), name)) {
-			continue;
-		}
-		free(name);
-		int n;
-		name = strdup(sdbkv_key(kv));
-		if (name && (arg && *arg)) {
-			if (!strcmp(arg, name)) {
-				match = true;
-			} else {
-				continue;
-			}
-		}
-		rz_cons_printf("%s %s {%s", sdbkv_value(kv), name, multiline ? "\n" : "");
-		char *p, *var = rz_str_newf("%s.%s", sdbkv_value(kv), name);
-		for (n = 0; (p = sdb_array_get(TDB, var, n, NULL)); n++) {
-			char *var2 = rz_str_newf("%s.%s", var, p);
-			if (var2) {
-				char *val = sdb_array_get(TDB, var2, 0, NULL);
-				if (val) {
-					char *arr = sdb_array_get(TDB, var2, 2, NULL);
-					int arrnum = atoi(arr);
-					free(arr);
-					if (multiline) {
-						rz_cons_printf("\t%s", val);
-						if (p && p[0] != '\0') {
-							rz_cons_printf("%s%s", strstr(val, " *") ? "" : " ", p);
-							if (arrnum) {
-								rz_cons_printf("[%d]", arrnum);
-							}
-						}
-						rz_cons_println(";");
-					} else {
-						rz_cons_printf("%s%s %s", space, val, p);
-						if (arrnum) {
-							rz_cons_printf("[%d]", arrnum);
-						}
-						rz_cons_print(";");
-						space = " ";
-					}
-					free(val);
-				}
-				free(var2);
-			}
-			free(p);
-		}
-		free(var);
-		rz_cons_println("};");
-		space = "";
-		if (match) {
-			break;
-		}
-	}
-	free(name);
-	ls_free(l);
-}
-
-static void print_enum_in_c_format(Sdb *TDB, const char *arg, bool multiline) {
-	char *name = NULL;
-	SdbKv *kv;
-	SdbListIter *iter;
-	SdbList *l = sdb_foreach_list(TDB, true);
-	const char *separator = "";
-	bool match = false;
-	ls_foreach (l, iter, kv) {
-		if (!strcmp(sdbkv_value(kv), "enum")) {
-			if (!name || strcmp(sdbkv_value(kv), name)) {
-				free(name);
-				name = strdup(sdbkv_key(kv));
-				if (name && (arg && *arg)) {
-					if (!strcmp(arg, name)) {
-						match = true;
-					} else {
-						continue;
-					}
-				}
-				rz_cons_printf("%s %s {%s", sdbkv_value(kv), name, multiline ? "\n" : "");
-				{
-					RzList *list = rz_type_get_enum(TDB, name);
-					if (list && !rz_list_empty(list)) {
-						RzListIter *iter;
-						RTypeEnum *member;
-						separator = multiline ? "\t" : "";
-						rz_list_foreach (list, iter, member) {
-							rz_cons_printf("%s%s = %" PFMT64u, separator, member->name, rz_num_math(NULL, member->val));
-							separator = multiline ? ",\n\t" : ", ";
-						}
-					}
-					rz_list_free(list);
-				}
-				rz_cons_println(multiline ? "\n};" : "};");
-				if (match) {
-					break;
-				}
-			}
-		}
-	}
-	free(name);
-	ls_free(l);
-}
-
-static bool printkey_cb(void *user, const char *k, const char *v) {
-	rz_cons_println(k);
-	return true;
-}
-
-static bool stdifunion(void *p, const char *k, const char *v) {
-	return !strncmp(v, "union", strlen("union") + 1);
-}
-
-static bool stdiftype(void *p, const char *k, const char *v) {
-	return !strncmp(v, "type", strlen("type") + 1);
-}
-
-static bool print_typelist_r_cb(void *p, const char *k, const char *v) {
-	rz_cons_printf("tk %s=%s\n", k, v);
-	return true;
-}
-
-static bool print_type_c(RzCore *core, const char *ctype) {
-	Sdb *TDB = core->analysis->sdb_types;
-	const char *type = rz_str_trim_head_ro(ctype);
-	const char *name = type ? strchr(type, '.') : NULL;
-	if (name && type) {
-		name++; // skip the '.'
-		if (rz_str_startswith(type, "struct")) {
-			print_struct_union_in_c_format(TDB, stdifstruct, name, true);
-		} else if (rz_str_startswith(type, "union")) {
-			print_struct_union_in_c_format(TDB, stdifunion, name, true);
-		} else if (rz_str_startswith(type, "enum")) {
-			print_enum_in_c_format(TDB, name, true);
-		} else if (rz_str_startswith(type, "typedef")) {
-			rz_core_list_typename_alias_c(core, name);
-		} else if (rz_str_startswith(type, "func")) {
-			types_function_print(core, name, RZ_OUTPUT_MODE_STANDARD, NULL);
-		}
-		return true;
-	}
-	return false;
-}
-
-static bool print_typelist_json_cb(void *p, const char *k, const char *v) {
-	RzCore *core = (RzCore *)p;
-	PJ *pj = pj_new();
-	pj_o(pj);
-	Sdb *sdb = core->analysis->sdb_types;
-	char *sizecmd = rz_str_newf("type.%s.size", k);
-	char *size_s = sdb_querys(sdb, NULL, -1, sizecmd);
-	char *formatcmd = rz_str_newf("type.%s", k);
-	char *format_s = sdb_querys(sdb, NULL, -1, formatcmd);
-	rz_str_trim(format_s);
-	pj_ks(pj, "type", k);
-	pj_ki(pj, "size", size_s ? atoi(size_s) : -1);
-	pj_ks(pj, "format", format_s);
-	pj_end(pj);
-	rz_cons_printf("%s", pj_string(pj));
-	pj_free(pj);
-	free(size_s);
-	free(format_s);
-	free(sizecmd);
-	free(formatcmd);
-	return true;
-}
-
-static void print_keys(Sdb *TDB, RzCore *core, SdbForeachCallback filter, SdbForeachCallback printfn_cb, bool json) {
-	SdbList *l = sdb_foreach_list_filter(TDB, filter, true);
-	SdbListIter *it;
-	SdbKv *kv;
-	const char *comma = "";
-
-	if (json) {
-		rz_cons_printf("[");
-	}
-	ls_foreach (l, it, kv) {
-		const char *k = sdbkv_key(kv);
-		if (!k || !*k) {
-			continue;
-		}
-		if (json) {
-			rz_cons_printf("%s", comma);
-			comma = ",";
-		}
-		printfn_cb(core, sdbkv_key(kv), sdbkv_value(kv));
-	}
-	if (json) {
-		rz_cons_printf("]\n");
-	}
-	ls_free(l);
-}
-
 static void typesList(RzCore *core, int mode) {
 	switch (mode) {
 	case 1:
 	case '*':
-		print_keys(core->analysis->sdb_types, core, NULL, print_typelist_r_cb, false);
+		rz_core_types_print_all(core, RZ_OUTPUT_MODE_RIZIN);
 		break;
 	case 'j':
-		print_keys(core->analysis->sdb_types, core, stdiftype, print_typelist_json_cb, true);
+		rz_core_types_print_all(core, RZ_OUTPUT_MODE_JSON);
 		break;
 	default:
-		print_keys(core->analysis->sdb_types, core, stdiftype, printkey_cb, false);
+		rz_core_types_print_all(core, RZ_OUTPUT_MODE_STANDARD);
 		break;
-	}
-}
-
-static void set_offset_hint(RzCore *core, RzAnalysisOp *op, const char *type, ut64 laddr, ut64 at, int offimm) {
-	char *res = rz_type_get_struct_memb(core->analysis->sdb_types, type, offimm);
-	const char *cmt = ((offimm == 0) && res) ? res : type;
-	if (offimm > 0) {
-		// set hint only if link is present
-		char *query = sdb_fmt("link.%08" PFMT64x, laddr);
-		if (res && sdb_const_get(core->analysis->sdb_types, query, 0)) {
-			rz_analysis_hint_set_offset(core->analysis, at, res);
-		}
-	} else if (cmt && rz_analysis_op_ismemref(op->type)) {
-		rz_meta_set_string(core->analysis, RZ_META_TYPE_VARTYPE, at, cmt);
-	}
-}
-
-static bool typedef_info(RzCore *core, const char *name) {
-	const char *istypedef;
-	Sdb *TDB = core->analysis->sdb_types;
-	istypedef = sdb_const_get(TDB, name, 0);
-	if (istypedef && !strncmp(istypedef, "typedef", 7)) {
-		const char *q = sdb_fmt("typedef.%s", name);
-		const char *res = sdb_const_get(TDB, q, 0);
-		if (res) {
-			rz_cons_println(res);
-		} else {
-			return false;
-		}
-	} else {
-		eprintf("This is not an typedef\n");
-		return false;
-	}
-	return true;
-}
-
-RZ_API void rz_core_list_loaded_typedefs(RzCore *core, RzOutputMode mode) {
-	PJ *pj = NULL;
-	Sdb *TDB = core->analysis->sdb_types;
-	if (mode == RZ_OUTPUT_MODE_JSON) {
-		pj = rz_core_pj_new(core);
-		if (!pj) {
-			return;
-		}
-		pj_o(pj);
-	}
-	char *name = NULL;
-	SdbKv *kv;
-	SdbListIter *iter;
-	SdbList *l = sdb_foreach_list(TDB, true);
-	ls_foreach (l, iter, kv) {
-		if (!strcmp(sdbkv_value(kv), "typedef")) {
-			if (!name || strcmp(sdbkv_value(kv), name)) {
-				free(name);
-				name = strdup(sdbkv_key(kv));
-				if (mode == RZ_OUTPUT_MODE_STANDARD) {
-					rz_cons_println(name);
-				} else {
-					const char *q = sdb_fmt("typedef.%s", name);
-					const char *res = sdb_const_get(TDB, q, 0);
-					pj_ks(pj, name, res);
-				}
-			}
-		}
-	}
-	if (mode == RZ_OUTPUT_MODE_JSON) {
-		pj_end(pj);
-		rz_cons_println(pj_string(pj));
-		pj_free(pj);
-	}
-	free(name);
-	ls_free(l);
-}
-
-RZ_API void rz_core_list_typename_alias_c(RzCore *core, const char *typedef_name) {
-	char *name = NULL;
-	SdbKv *kv;
-	SdbListIter *iter;
-	Sdb *TDB = core->analysis->sdb_types;
-	SdbList *l = sdb_foreach_list(TDB, true);
-	bool match = false;
-	ls_foreach (l, iter, kv) {
-		if (!strcmp(sdbkv_value(kv), "typedef")) {
-			if (!name || strcmp(sdbkv_value(kv), name)) {
-				free(name);
-				name = strdup(sdbkv_key(kv));
-				if (name && (typedef_name && *typedef_name)) {
-					if (!strcmp(typedef_name, name)) {
-						match = true;
-					} else {
-						continue;
-					}
-				}
-				const char *q = sdb_fmt("typedef.%s", name);
-				const char *res = sdb_const_get(TDB, q, 0);
-				if (res) {
-					rz_cons_printf("%s %s %s;\n", sdbkv_value(kv), res, name);
-				}
-				if (match) {
-					break;
-				}
-			}
-		}
-	}
-	free(name);
-	ls_free(l);
-}
-
-RZ_API int rz_core_get_stacksz(RzCore *core, ut64 from, ut64 to) {
-	int stack = 0, maxstack = 0;
-	ut64 at = from;
-
-	if (from >= to) {
-		return 0;
-	}
-	const int mininstrsz = rz_analysis_archinfo(core->analysis, RZ_ANALYSIS_ARCHINFO_MIN_OP_SIZE);
-	const int minopcode = RZ_MAX(1, mininstrsz);
-	while (at < to) {
-		RzAnalysisOp *op = rz_core_analysis_op(core, at, RZ_ANALYSIS_OP_MASK_BASIC);
-		if (!op || op->size <= 0) {
-			at += minopcode;
-			continue;
-		}
-		if ((op->stackop == RZ_ANALYSIS_STACK_INC) && RZ_ABS(op->stackptr) < 8096) {
-			stack += op->stackptr;
-			if (stack > maxstack) {
-				maxstack = stack;
-			}
-		}
-		at += op->size;
-		rz_analysis_op_free(op);
-	}
-	return maxstack;
-}
-
-static void set_retval(RzCore *core, ut64 at) {
-	RzAnalysis *analysis = core->analysis;
-	RzAnalysisHint *hint = rz_analysis_hint_get(analysis, at);
-	RzAnalysisFunction *fcn = rz_analysis_get_fcn_in(analysis, at, 0);
-
-	if (!hint || !fcn || !fcn->name) {
-		goto beach;
-	}
-	if (hint->ret == UT64_MAX) {
-		goto beach;
-	}
-	const char *cc = rz_analysis_cc_func(core->analysis, fcn->name);
-	const char *regname = rz_analysis_cc_ret(analysis, cc);
-	if (regname) {
-		RzRegItem *reg = rz_reg_get(analysis->reg, regname, -1);
-		if (reg) {
-			rz_reg_set_value(analysis->reg, reg, hint->ret);
-		}
-	}
-beach:
-	rz_analysis_hint_free(hint);
-	return;
-}
-
-RZ_API void rz_core_link_stroff(RzCore *core, RzAnalysisFunction *fcn) {
-	RzAnalysisBlock *bb;
-	RzListIter *it;
-	RzAnalysisOp aop = { 0 };
-	bool ioCache = rz_config_get_i(core->config, "io.cache");
-	bool stack_set = false;
-	bool resolved = false;
-	const char *varpfx;
-	int dbg_follow = rz_config_get_i(core->config, "dbg.follow");
-	Sdb *TDB = core->analysis->sdb_types;
-	RzAnalysisEsil *esil;
-	int iotrap = rz_config_get_i(core->config, "esil.iotrap");
-	int stacksize = rz_config_get_i(core->config, "esil.stack.depth");
-	unsigned int addrsize = rz_config_get_i(core->config, "esil.addr.size");
-	const char *pc_name = rz_reg_get_name(core->analysis->reg, RZ_REG_NAME_PC);
-	const char *sp_name = rz_reg_get_name(core->analysis->reg, RZ_REG_NAME_SP);
-	RzRegItem *pc = rz_reg_get(core->analysis->reg, pc_name, -1);
-
-	if (!fcn) {
-		return;
-	}
-	if (!(esil = rz_analysis_esil_new(stacksize, iotrap, addrsize))) {
-		return;
-	}
-	rz_analysis_esil_setup(esil, core->analysis, 0, 0, 0);
-	int i, ret, bsize = RZ_MAX(64, core->blocksize);
-	const int mininstrsz = rz_analysis_archinfo(core->analysis, RZ_ANALYSIS_ARCHINFO_MIN_OP_SIZE);
-	const int maxinstrsz = rz_analysis_archinfo(core->analysis, RZ_ANALYSIS_ARCHINFO_MAX_OP_SIZE);
-	const int minopcode = RZ_MAX(1, mininstrsz);
-	ut8 *buf = malloc(bsize);
-	if (!buf) {
-		free(buf);
-		rz_analysis_esil_free(esil);
-		return;
-	}
-	rz_reg_arena_push(core->analysis->reg);
-	rz_debug_reg_sync(core->dbg, RZ_REG_TYPE_ALL, true);
-	ut64 spval = rz_reg_getv(esil->analysis->reg, sp_name);
-	if (spval) {
-		// reset stack pointer to initial value
-		RzRegItem *sp = rz_reg_get(esil->analysis->reg, sp_name, -1);
-		ut64 curpc = rz_reg_getv(esil->analysis->reg, pc_name);
-		int stacksz = rz_core_get_stacksz(core, fcn->addr, curpc);
-		if (stacksz > 0) {
-			rz_reg_arena_zero(esil->analysis->reg); // clear prev reg values
-			rz_reg_set_value(esil->analysis->reg, sp, spval + stacksz);
-		}
-	} else {
-		// initialize stack
-		rz_core_analysis_esil_init_mem(core, NULL, UT64_MAX, UT32_MAX);
-		stack_set = true;
-	}
-	rz_config_set_i(core->config, "io.cache", 1);
-	rz_config_set_i(core->config, "dbg.follow", 0);
-	ut64 oldoff = core->offset;
-	rz_cons_break_push(NULL, NULL);
-	// TODO: The algorithm can be more accurate if blocks are followed by their jmp/fail, not just by address
-	rz_list_sort(fcn->bbs, bb_cmpaddr);
-	rz_list_foreach (fcn->bbs, it, bb) {
-		ut64 at = bb->addr;
-		ut64 to = bb->addr + bb->size;
-		rz_reg_set_value(esil->analysis->reg, pc, at);
-		for (i = 0; at < to; i++) {
-			if (rz_cons_is_breaked()) {
-				goto beach;
-			}
-			if (at < bb->addr) {
-				break;
-			}
-			if (i >= (bsize - maxinstrsz)) {
-				i = 0;
-			}
-			if (!i) {
-				rz_io_read_at(core->io, at, buf, bsize);
-			}
-			ret = rz_analysis_op(core->analysis, &aop, at, buf + i, bsize - i, RZ_ANALYSIS_OP_MASK_VAL);
-			if (ret <= 0) {
-				i += minopcode;
-				at += minopcode;
-				rz_analysis_op_fini(&aop);
-				continue;
-			}
-			i += ret - 1;
-			at += ret;
-			int index = 0;
-			if (aop.ireg) {
-				index = rz_reg_getv(esil->analysis->reg, aop.ireg) * aop.scale;
-			}
-			int j, src_imm = -1, dst_imm = -1;
-			ut64 src_addr = UT64_MAX;
-			ut64 dst_addr = UT64_MAX;
-			for (j = 0; j < 3; j++) {
-				if (aop.src[j] && aop.src[j]->reg && aop.src[j]->reg->name) {
-					src_addr = rz_reg_getv(esil->analysis->reg, aop.src[j]->reg->name) + index;
-					src_imm = aop.src[j]->delta;
-				}
-			}
-			if (aop.dst && aop.dst->reg && aop.dst->reg->name) {
-				dst_addr = rz_reg_getv(esil->analysis->reg, aop.dst->reg->name) + index;
-				dst_imm = aop.dst->delta;
-			}
-			RzAnalysisVar *var = rz_analysis_get_used_function_var(core->analysis, aop.addr);
-			if (false) { // src_addr != UT64_MAX || dst_addr != UT64_MAX) {
-				//  if (src_addr == UT64_MAX && dst_addr == UT64_MAX) {
-				rz_analysis_op_fini(&aop);
-				continue;
-			}
-			char *slink = rz_type_link_at(TDB, src_addr);
-			char *vlink = rz_type_link_at(TDB, src_addr + src_imm);
-			char *dlink = rz_type_link_at(TDB, dst_addr);
-			//TODO: Handle register based arg for struct offset propgation
-			if (vlink && var && var->kind != 'r') {
-				if (rz_type_kind(TDB, vlink) == RZ_TYPE_UNION) {
-					varpfx = "union";
-				} else {
-					varpfx = "struct";
-				}
-				// if a var addr matches with struct , change it's type and name
-				// var int local_e0h --> var struct foo
-				if (strcmp(var->name, vlink) && !resolved) {
-					resolved = true;
-					rz_analysis_var_set_type(var, varpfx);
-					rz_analysis_var_rename(var, vlink, false);
-				}
-			} else if (slink) {
-				set_offset_hint(core, &aop, slink, src_addr, at - ret, src_imm);
-			} else if (dlink) {
-				set_offset_hint(core, &aop, dlink, dst_addr, at - ret, dst_imm);
-			}
-			if (rz_analysis_op_nonlinear(aop.type)) {
-				rz_reg_set_value(esil->analysis->reg, pc, at);
-				set_retval(core, at - ret);
-			} else {
-				rz_core_esil_step(core, UT64_MAX, NULL, NULL, false);
-			}
-			free(dlink);
-			free(vlink);
-			free(slink);
-			rz_analysis_op_fini(&aop);
-		}
-	}
-beach:
-	rz_io_cache_reset(core->io, core->io->cached); // drop cache writes
-	rz_config_set_i(core->config, "io.cache", ioCache);
-	rz_config_set_i(core->config, "dbg.follow", dbg_follow);
-	if (stack_set) {
-		rz_core_analysis_esil_init_mem_del(core, NULL, UT64_MAX, UT32_MAX);
-	}
-	rz_core_seek(core, oldoff, true);
-	rz_analysis_esil_free(esil);
-	rz_reg_arena_pop(core->analysis->reg);
-	rz_core_regs2flags(core);
-	rz_cons_break_pop();
-	free(buf);
-}
-
-static void print_all_format(RzCore *core, Sdb *TDB, SdbForeachCallback sdbcb) {
-	SdbList *l = sdb_foreach_list_filter(TDB, sdbcb, true);
-	SdbListIter *it;
-	SdbKv *kv;
-	ls_foreach (l, it, kv) {
-		type_show_format(core, sdbkv_key(kv), RZ_OUTPUT_MODE_RIZIN);
-	}
-	ls_free(l);
-}
-
-static void print_all_struct_format(RzCore *core, Sdb *TDB) {
-	print_all_format(core, TDB, stdifstruct);
-}
-
-static void print_all_union_format(RzCore *core, Sdb *TDB) {
-	print_all_format(core, TDB, stdifunion);
-}
-
-static void type_list_c_all(RzCore *core) {
-	Sdb *TDB = core->analysis->sdb_types;
-	//types_function_print_all(core, RZ_OUTPUT_MODE_STANDARD);
-	// List all unions in the C format with newlines
-	print_struct_union_in_c_format(TDB, stdifunion, NULL, true);
-	// List all structures in the C format with newlines
-	rz_core_cmd0(core, "tsc");
-	// List all typedefs in the C format with newlines
-	rz_core_list_typename_alias_c(core, NULL);
-	// List all enums in the C format with newlines
-	print_enum_in_c_format(TDB, NULL, true);
-}
-
-static void type_list_c_all_nl(RzCore *core) {
-	Sdb *TDB = core->analysis->sdb_types;
-	print_struct_union_in_c_format(TDB, stdifunion, NULL, false);
-	rz_core_cmd0(core, "tsd");
-	rz_core_list_typename_alias_c(core, NULL);
-	print_enum_in_c_format(TDB, NULL, false);
-}
-
-static RzCmdStatus type_format_print(RzCore *core, const char *type, ut64 address) {
-	Sdb *TDB = core->analysis->sdb_types;
-	char *fmt = rz_type_format(TDB, type);
-	if (!fmt || !*fmt) {
-		eprintf("Cannot find type %s\n", type);
-		return RZ_CMD_STATUS_ERROR;
-	}
-	rz_core_cmdf(core, "pf %s @ 0x%08" PFMT64x "\n", fmt, address);
-	return RZ_CMD_STATUS_OK;
-}
-
-static RzCmdStatus type_format_print_variable(RzCore *core, const char *type, const char *varname) {
-	Sdb *TDB = core->analysis->sdb_types;
-	char *fmt = rz_type_format(TDB, type);
-	if (!fmt || !*fmt) {
-		eprintf("Cannot find type \"%s\"\n", type);
-		return RZ_CMD_STATUS_ERROR;
-	}
-	RzAnalysisFunction *fcn = rz_analysis_get_fcn_in(core->analysis, core->offset, -1);
-	if (!fcn) {
-		eprintf("Cannot find function at the current offset\n");
-		return RZ_CMD_STATUS_ERROR;
-	}
-	RzAnalysisVar *var = rz_analysis_function_get_var_byname(fcn, varname);
-	if (!var) {
-		eprintf("Cannot find variable \"%s\" in the current function\n", varname);
-		return RZ_CMD_STATUS_ERROR;
-	}
-	ut64 addr = rz_analysis_var_addr(var);
-	rz_core_cmdf(core, "pf %s @ 0x%08" PFMT64x "\n", fmt, addr);
-	return RZ_CMD_STATUS_OK;
-}
-
-static RzCmdStatus type_format_print_value(RzCore *core, const char *type, ut64 val) {
-	Sdb *TDB = core->analysis->sdb_types;
-	char *fmt = rz_type_format(TDB, type);
-	if (!fmt || !*fmt) {
-		eprintf("Cannot find type %s\n", type);
-		return RZ_CMD_STATUS_ERROR;
-	}
-	rz_core_cmdf(core, "pf %s @v:0x%08" PFMT64x "\n", fmt, val);
-	return RZ_CMD_STATUS_OK;
-}
-
-static RzCmdStatus type_format_print_hexstring(RzCore *core, const char *type, const char *hexpairs) {
-	Sdb *TDB = core->analysis->sdb_types;
-	char *fmt = rz_type_format(TDB, type);
-	if (!fmt && !*fmt) {
-		eprintf("Cannot find type %s\n", type);
-		return RZ_CMD_STATUS_ERROR;
-	}
-	rz_core_cmdf(core, "pf %s @x:%s", fmt, hexpairs);
-	return RZ_CMD_STATUS_OK;
-}
-
-static void type_define(RzCore *core, const char *type) {
-	// Add trailing semicolon to force the valid C syntax
-	// It allows us to skip the trailing semicolon in the input
-	// to reduce the unnecessary typing
-	char *tmp = rz_str_newf("%s;", type);
-	if (!tmp) {
-		return;
-	}
-	char *error_msg = NULL;
-	char *out = rz_parse_c_string(core->analysis, tmp, &error_msg);
-	free(tmp);
-	if (out) {
-		rz_analysis_save_parsed_type(core->analysis, out);
-		free(out);
-	}
-	if (error_msg) {
-		eprintf("%s", error_msg);
-		free(error_msg);
-	}
-}
-
-static void types_open_file(RzCore *core, const char *path) {
-	const char *dir = rz_config_get(core->config, "dir.types");
-	char *homefile = NULL;
-	if (*path == '~') {
-		if (path[1] && path[2]) {
-			homefile = rz_str_home(path + 2);
-			path = homefile;
-		}
-	}
-	if (!strcmp(path, "-")) {
-		char *tmp = rz_core_editor(core, "*.h", "");
-		if (tmp) {
-			char *error_msg = NULL;
-			char *out = rz_parse_c_string(core->analysis, tmp, &error_msg);
-			if (out) {
-				rz_analysis_save_parsed_type(core->analysis, out);
-				free(out);
-			}
-			if (error_msg) {
-				fprintf(stderr, "%s", error_msg);
-				free(error_msg);
-			}
-			free(tmp);
-		}
-	} else {
-		char *error_msg = NULL;
-		char *out = rz_parse_c_file(core->analysis, path, dir, &error_msg);
-		if (out) {
-			rz_analysis_save_parsed_type(core->analysis, out);
-			free(out);
-		}
-		if (error_msg) {
-			fprintf(stderr, "%s", error_msg);
-			free(error_msg);
-		}
-	}
-	free(homefile);
-}
-
-static void types_open_editor(RzCore *core, const char *typename) {
-	Sdb *TDB = core->analysis->sdb_types;
-	char *str = rz_core_cmd_strf(core, "tc %s", typename ? typename : "");
-	char *tmp = rz_core_editor(core, "*.h", str);
-	if (tmp) {
-		char *error_msg = NULL;
-		char *out = rz_parse_c_string(core->analysis, tmp, &error_msg);
-		if (out) {
-			// remove previous types and save new edited types
-			sdb_reset(TDB);
-			rz_parse_c_reset(core->parser);
-			rz_analysis_save_parsed_type(core->analysis, out);
-			free(out);
-		}
-		if (error_msg) {
-			eprintf("%s\n", error_msg);
-			free(error_msg);
-		}
-		free(tmp);
-	}
-	free(str);
-}
-
-static void types_open_sdb(RzCore *core, const char *path) {
-	Sdb *TDB = core->analysis->sdb_types;
-	if (rz_file_exists(path)) {
-		Sdb *db_tmp = sdb_new(0, path, 0);
-		sdb_merge(TDB, db_tmp);
-		sdb_close(db_tmp);
-		sdb_free(db_tmp);
 	}
 }
 
@@ -1445,7 +554,6 @@ RZ_IPI int rz_cmd_type(void *data, const char *input) {
 	RzCore *core = (RzCore *)data;
 	Sdb *TDB = core->analysis->sdb_types;
 	char *res;
-	TDB_ = TDB; // HACK
 
 	switch (input[0]) {
 	case 'n': // "tn"
@@ -1459,30 +567,30 @@ RZ_IPI int rz_cmd_type(void *data, const char *input) {
 			break;
 		case '*':
 			if (input[2] == ' ') {
-				type_show_format(core, rz_str_trim_head_ro(input + 2), RZ_OUTPUT_MODE_RIZIN);
+				rz_core_types_show_format(core, rz_str_trim_head_ro(input + 2), RZ_OUTPUT_MODE_RIZIN);
 			} else {
-				print_all_union_format(core, TDB);
+				rz_core_types_union_print_format_all(core, TDB);
 			}
 			break;
 		case 'j': // "tuj"
 			if (input[2]) {
-				type_show_format(core, rz_str_trim_head_ro(input + 2), RZ_OUTPUT_MODE_JSON);
+				rz_core_types_show_format(core, rz_str_trim_head_ro(input + 2), RZ_OUTPUT_MODE_JSON);
 				rz_cons_newline();
 			} else {
-				print_struct_union_list_json(TDB, stdifunion);
+				rz_types_union_print_json(TDB);
 			}
 			break;
 		case 'c': // "tuc"
-			print_struct_union_in_c_format(TDB, stdifunion, rz_str_trim_head_ro(input + 2), true);
+			rz_types_union_print_c(TDB, rz_str_trim_head_ro(input + 2), true);
 			break;
 		case 'd': // "tud"
-			print_struct_union_in_c_format(TDB, stdifunion, rz_str_trim_head_ro(input + 2), false);
+			rz_types_union_print_c(TDB, rz_str_trim_head_ro(input + 2), false);
 			break;
 		case ' ': // "tu "
-			type_show_format(core, rz_str_trim_head_ro(input + 1), RZ_OUTPUT_MODE_STANDARD);
+			rz_core_types_show_format(core, rz_str_trim_head_ro(input + 1), RZ_OUTPUT_MODE_STANDARD);
 			break;
 		case 0:
-			print_keys(TDB, core, stdifunion, printkey_cb, false);
+			rz_types_union_print_sdb(TDB);
 			break;
 		}
 	} break;
@@ -1527,13 +635,13 @@ RZ_IPI int rz_cmd_type(void *data, const char *input) {
 			break;
 		case '*':
 			if (input[2] == ' ') {
-				type_show_format(core, rz_str_trim_head_ro(input + 2), RZ_OUTPUT_MODE_RIZIN);
+				rz_core_types_show_format(core, rz_str_trim_head_ro(input + 1), RZ_OUTPUT_MODE_RIZIN);
 			} else {
-				print_all_struct_format(core, TDB);
+				rz_core_types_struct_print_format_all(core, TDB);
 			}
 			break;
 		case ' ':
-			type_show_format(core, rz_str_trim_head_ro(input + 1), RZ_OUTPUT_MODE_STANDARD);
+			rz_core_types_show_format(core, rz_str_trim_head_ro(input + 1), RZ_OUTPUT_MODE_STANDARD);
 			break;
 		case 's':
 			if (input[2] == ' ') {
@@ -1543,21 +651,21 @@ RZ_IPI int rz_cmd_type(void *data, const char *input) {
 			}
 			break;
 		case 0:
-			print_keys(TDB, core, stdifstruct, printkey_cb, false);
+			rz_types_struct_print_sdb(TDB);
 			break;
 		case 'c': // "tsc"
-			print_struct_union_in_c_format(TDB, stdifstruct, rz_str_trim_head_ro(input + 2), true);
+			rz_types_struct_print_c(TDB, rz_str_trim_head_ro(input + 2), true);
 			break;
 		case 'd': // "tsd"
-			print_struct_union_in_c_format(TDB, stdifstruct, rz_str_trim_head_ro(input + 2), false);
+			rz_types_struct_print_c(TDB, rz_str_trim_head_ro(input + 2), false);
 			break;
 		case 'j': // "tsj"
 			// TODO: current output is a bit poor, will be good to improve
 			if (input[2]) {
-				type_show_format(core, rz_str_trim_head_ro(input + 2), RZ_OUTPUT_MODE_JSON);
+				rz_core_types_show_format(core, rz_str_trim_head_ro(input + 2), RZ_OUTPUT_MODE_JSON);
 				rz_cons_newline();
 			} else {
-				print_struct_union_list_json(TDB, stdifstruct);
+				rz_types_struct_print_json(TDB);
 			}
 			break;
 		}
@@ -1582,13 +690,13 @@ RZ_IPI int rz_cmd_type(void *data, const char *input) {
 			break;
 		case 'j': // "tej"
 			if (input[2] == 0) { // "tej"
-				types_enum_print_all(core, RZ_OUTPUT_MODE_JSON);
+				rz_core_types_enum_print_all(core, RZ_OUTPUT_MODE_JSON);
 			} else { // "tej ENUM"
 				if (member_value) {
 					types_enum_member_find(core, name, member_value);
 				} else {
 					PJ *pj = rz_core_pj_new(core);
-					types_enum_print(core, name, RZ_OUTPUT_MODE_JSON, pj);
+					rz_core_types_enum_print(core, name, RZ_OUTPUT_MODE_JSON, pj);
 					pj_end(pj);
 					rz_cons_println(pj_string(pj));
 					pj_free(pj);
@@ -1597,19 +705,19 @@ RZ_IPI int rz_cmd_type(void *data, const char *input) {
 			break;
 		case 'k': // "tek"
 			if (input[2] == 0) { // "tek"
-				types_enum_print_all(core, RZ_OUTPUT_MODE_SDB);
+				rz_core_types_enum_print_all(core, RZ_OUTPUT_MODE_SDB);
 			} else { // "tek ENUM"
-				types_enum_print(core, name, RZ_OUTPUT_MODE_SDB, NULL);
+				rz_core_types_enum_print(core, name, RZ_OUTPUT_MODE_SDB, NULL);
 			}
 			break;
 		case 'b': // "teb"
 			res = rz_type_enum_member(TDB, name, member_value, 0);
 			break;
 		case 'c': // "tec"
-			print_enum_in_c_format(TDB, rz_str_trim_head_ro(input + 2), true);
+			rz_types_enum_print_c(TDB, rz_str_trim_head_ro(input + 2), true);
 			break;
 		case 'd': // "ted"
-			print_enum_in_c_format(TDB, rz_str_trim_head_ro(input + 2), false);
+			rz_types_enum_print_c(TDB, rz_str_trim_head_ro(input + 2), false);
 			break;
 		case 'f': // "tef"
 			if (member_value) {
@@ -1620,11 +728,11 @@ RZ_IPI int rz_cmd_type(void *data, const char *input) {
 			if (member_value) {
 				types_enum_member_find(core, name, member_value);
 			} else {
-				types_enum_print(core, name, RZ_OUTPUT_MODE_STANDARD, NULL);
+				rz_core_types_enum_print(core, name, RZ_OUTPUT_MODE_STANDARD, NULL);
 			}
 			break;
 		case '\0': {
-			types_enum_print_all(core, RZ_OUTPUT_MODE_QUIET);
+			rz_core_types_enum_print_all(core, RZ_OUTPUT_MODE_QUIET);
 		} break;
 		}
 		free(name);
@@ -1635,7 +743,7 @@ RZ_IPI int rz_cmd_type(void *data, const char *input) {
 		}
 	} break;
 	case ' ':
-		type_show_format(core, input + 1, RZ_OUTPUT_MODE_STANDARD);
+		rz_core_types_show_format(core, input + 1, RZ_OUTPUT_MODE_STANDARD);
 		break;
 	// t* - list all types in 'pf' syntax
 	case 'j': // "tj"
@@ -1787,14 +895,14 @@ RZ_IPI int rz_cmd_type(void *data, const char *input) {
 				}
 			}
 			rz_str_trim(type);
-			types_link(core, type, addr);
+			rz_core_types_link(core, type, addr);
 			free(type);
 			break;
 		}
 		case 's': {
 			char *ptr = rz_str_trim_dup(input + 2);
 			ut64 addr = rz_num_math(core->num, ptr);
-			types_link_show(core, addr);
+			rz_core_types_link_show(core, addr);
 			free(ptr);
 			break;
 		}
@@ -1812,16 +920,16 @@ RZ_IPI int rz_cmd_type(void *data, const char *input) {
 			}
 			break;
 		case '*':
-			types_link_print_all(core, RZ_OUTPUT_MODE_RIZIN);
+			rz_core_types_link_print_all(core, RZ_OUTPUT_MODE_RIZIN);
 			break;
 		case 'l':
-			types_link_print_all(core, RZ_OUTPUT_MODE_LONG);
+			rz_core_types_link_print_all(core, RZ_OUTPUT_MODE_LONG);
 			break;
 		case 'j':
-			types_link_print_all(core, RZ_OUTPUT_MODE_JSON);
+			rz_core_types_link_print_all(core, RZ_OUTPUT_MODE_JSON);
 			break;
 		case '\0':
-			types_link_print_all(core, RZ_OUTPUT_MODE_STANDARD);
+			rz_core_types_link_print_all(core, RZ_OUTPUT_MODE_STANDARD);
 			break;
 		}
 		break;
@@ -1886,31 +994,31 @@ RZ_IPI int rz_cmd_type(void *data, const char *input) {
 	case 'f': // "tf"
 		switch (input[1]) {
 		case 0: // "tf"
-			types_function_print_all(core, RZ_OUTPUT_MODE_STANDARD);
+			rz_core_types_function_print_all(core, RZ_OUTPUT_MODE_STANDARD);
 			break;
 		case 'k': // "tfk"
 			if (input[2] == ' ') {
 				const char *name = rz_str_trim_head_ro(input + 3);
-				types_function_print(core, name, RZ_OUTPUT_MODE_SDB, NULL);
+				rz_types_function_print(TDB, name, RZ_OUTPUT_MODE_SDB, NULL);
 			} else {
-				types_function_print_all(core, RZ_OUTPUT_MODE_SDB);
+				rz_core_types_function_print_all(core, RZ_OUTPUT_MODE_SDB);
 			}
 			break;
 		case 'j': // "tfj"
 			if (input[2] == ' ') {
 				const char *name = rz_str_trim_head_ro(input + 2);
 				PJ *pj = rz_core_pj_new(core);
-				types_function_print(core, name, RZ_OUTPUT_MODE_JSON, pj);
+				rz_types_function_print(TDB, name, RZ_OUTPUT_MODE_JSON, pj);
 				pj_end(pj);
 				rz_cons_println(pj_string(pj));
 				pj_free(pj);
 			} else {
-				types_function_print_all(core, RZ_OUTPUT_MODE_JSON);
+				rz_core_types_function_print_all(core, RZ_OUTPUT_MODE_JSON);
 			}
 			break;
 		case ' ': {
 			const char *name = rz_str_trim_head_ro(input + 2);
-			types_function_print(core, name, RZ_OUTPUT_MODE_SDB, NULL);
+			rz_types_function_print(TDB, name, RZ_OUTPUT_MODE_SDB, NULL);
 			break;
 		}
 		default:
@@ -1926,7 +1034,7 @@ RZ_IPI int rz_cmd_type(void *data, const char *input) {
 			break;
 		}
 		if (input[1] == 'c') { // "ttc"
-			rz_core_list_typename_alias_c(core, input + 2);
+			rz_types_typedef_print_c(TDB, input + 2);
 			break;
 		}
 		if (input[1] == '?') { // "tt?"
@@ -1938,7 +1046,7 @@ RZ_IPI int rz_cmd_type(void *data, const char *input) {
 			break;
 		}
 		char *s = strdup(input + 2);
-		typedef_info(core, s);
+		rz_core_types_typedef_info(core, s);
 		free(s);
 	} break;
 	case '?':
@@ -1948,12 +1056,15 @@ RZ_IPI int rz_cmd_type(void *data, const char *input) {
 	return true;
 }
 
+// =============================================================================
+//                        END    DEPRECATED
+
 RZ_IPI RzCmdStatus rz_type_cc_list_handler(RzCore *core, int argc, const char **argv, RzOutputMode mode) {
 	const char *cc = argc > 1 ? argv[1] : NULL;
 	if (cc) {
 		types_cc_print(core, cc, mode);
 	} else {
-		types_cc_print_all(core, mode);
+		rz_core_types_calling_conventions_print(core, mode);
 	}
 	return RZ_CMD_STATUS_OK;
 }
@@ -2002,7 +1113,7 @@ RZ_IPI RzCmdStatus rz_type_list_enum_handler(RzCore *core, int argc, const char 
 			return types_enum_member_find(core, enum_name, member_value);
 		} else {
 			PJ *pj = (mode == RZ_OUTPUT_MODE_JSON) ? rz_core_pj_new(core) : NULL;
-			types_enum_print(core, enum_name, mode, pj);
+			rz_core_types_enum_print(core, enum_name, mode, pj);
 			if (mode == RZ_OUTPUT_MODE_JSON) {
 				rz_cons_println(pj_string(pj));
 				pj_free(pj);
@@ -2013,7 +1124,7 @@ RZ_IPI RzCmdStatus rz_type_list_enum_handler(RzCore *core, int argc, const char 
 		if (mode == RZ_OUTPUT_MODE_STANDARD) {
 			mode = RZ_OUTPUT_MODE_QUIET;
 		}
-		types_enum_print_all(core, mode);
+		rz_core_types_enum_print_all(core, mode);
 	}
 	return RZ_CMD_STATUS_OK;
 }
@@ -2035,14 +1146,14 @@ RZ_IPI RzCmdStatus rz_type_enum_bitfield_handler(RzCore *core, int argc, const c
 RZ_IPI RzCmdStatus rz_type_enum_c_handler(RzCore *core, int argc, const char **argv) {
 	const char *enum_name = argc > 1 ? argv[1] : NULL;
 	Sdb *TDB = core->analysis->sdb_types;
-	print_enum_in_c_format(TDB, enum_name, true);
+	rz_types_enum_print_c(TDB, enum_name, true);
 	return RZ_CMD_STATUS_OK;
 }
 
 RZ_IPI RzCmdStatus rz_type_enum_c_nl_handler(RzCore *core, int argc, const char **argv) {
 	const char *enum_name = argc > 1 ? argv[1] : NULL;
 	Sdb *TDB = core->analysis->sdb_types;
-	print_enum_in_c_format(TDB, enum_name, false);
+	rz_types_enum_print_c(TDB, enum_name, false);
 	return RZ_CMD_STATUS_OK;
 }
 
@@ -2055,13 +1166,13 @@ RZ_IPI RzCmdStatus rz_type_list_function_handler(RzCore *core, int argc, const c
 	const char *function = argc > 1 ? argv[1] : NULL;
 	if (function) {
 		PJ *pj = (mode == RZ_OUTPUT_MODE_JSON) ? rz_core_pj_new(core) : NULL;
-		types_function_print(core, function, mode, pj);
+		rz_types_function_print(core->analysis->sdb_types, function, mode, pj);
 		if (mode == RZ_OUTPUT_MODE_JSON) {
 			rz_cons_println(pj_string(pj));
 			pj_free(pj);
 		}
 	} else {
-		types_function_print_all(core, mode);
+		rz_core_types_function_print_all(core, mode);
 	}
 	return RZ_CMD_STATUS_OK;
 }
@@ -2088,16 +1199,16 @@ RZ_IPI RzCmdStatus rz_type_link_handler(RzCore *core, int argc, const char **arg
 	const char *name = argc > 1 ? argv[1] : NULL;
 	ut64 addr = argc > 2 ? rz_num_math(core->num, argv[2]) : core->offset;
 	if (name) {
-		types_link(core, name, addr);
+		rz_core_types_link(core, name, addr);
 	} else {
-		types_link_print_all(core, mode);
+		rz_core_types_link_print_all(core, mode);
 	}
 	return RZ_CMD_STATUS_OK;
 }
 
 RZ_IPI RzCmdStatus rz_type_link_show_handler(RzCore *core, int argc, const char **argv) {
 	ut64 addr = rz_num_math(core->num, argv[1]);
-	types_link_show(core, addr);
+	rz_core_types_link_show(core, addr);
 	return RZ_CMD_STATUS_OK;
 }
 
@@ -2124,7 +1235,7 @@ RZ_IPI RzCmdStatus rz_type_list_noreturn_handler(RzCore *core, int argc, const c
 			rz_analysis_noreturn_add(core->analysis, name, UT64_MAX);
 		}
 	} else {
-		rz_core_analysis_noreturn_print(core, mode);
+		rz_core_types_function_noreturn_print(core, mode);
 	}
 	return RZ_CMD_STATUS_OK;
 }
@@ -2137,7 +1248,7 @@ RZ_IPI RzCmdStatus rz_type_noreturn_del_handler(RzCore *core, int argc, const ch
 }
 
 RZ_IPI RzCmdStatus rz_type_noreturn_del_all_handler(RzCore *core, int argc, const char **argv) {
-	RzList *noretl = rz_core_analysis_noreturn(core);
+	RzList *noretl = rz_types_function_noreturn(core->analysis->sdb_types);
 	RzListIter *iter;
 	char *name;
 	rz_list_foreach (noretl, iter, name) {
@@ -2190,14 +1301,14 @@ RZ_IPI RzCmdStatus rz_type_list_structure_handler(RzCore *core, int argc, const 
 	const char *typename = argc > 1 ? argv[1] : NULL;
 	Sdb *TDB = core->analysis->sdb_types;
 	if (typename) {
-		type_show_format(core, typename, mode);
+		rz_core_types_show_format(core, typename, mode);
 	} else {
 		if (mode == RZ_OUTPUT_MODE_RIZIN) {
-			print_all_struct_format(core, TDB);
+			rz_core_types_struct_print_format_all(core, TDB);
 		} else if (mode == RZ_OUTPUT_MODE_JSON) {
-			print_struct_union_list_json(TDB, stdifstruct);
+			rz_types_struct_print_json(TDB);
 		} else {
-			print_keys(TDB, core, stdifstruct, printkey_cb, false);
+			rz_types_struct_print_sdb(TDB);
 		}
 	}
 	return RZ_CMD_STATUS_OK;
@@ -2206,21 +1317,21 @@ RZ_IPI RzCmdStatus rz_type_list_structure_handler(RzCore *core, int argc, const 
 RZ_IPI RzCmdStatus rz_type_structure_c_handler(RzCore *core, int argc, const char **argv) {
 	const char *typename = argc > 1 ? argv[1] : NULL;
 	Sdb *TDB = core->analysis->sdb_types;
-	print_struct_union_in_c_format(TDB, stdifstruct, typename, true);
+	rz_types_struct_print_c(TDB, typename, true);
 	return RZ_CMD_STATUS_OK;
 }
 
 RZ_IPI RzCmdStatus rz_type_structure_c_nl_handler(RzCore *core, int argc, const char **argv) {
 	const char *typename = argc > 1 ? argv[1] : NULL;
 	Sdb *TDB = core->analysis->sdb_types;
-	print_struct_union_in_c_format(TDB, stdifstruct, typename, false);
+	rz_types_struct_print_c(TDB, typename, false);
 	return RZ_CMD_STATUS_OK;
 }
 
 RZ_IPI RzCmdStatus rz_type_list_typedef_handler(RzCore *core, int argc, const char **argv, RzOutputMode mode) {
 	const char *typename = argc > 1 ? argv[1] : NULL;
 	if (typename) {
-		if (!typedef_info(core, typename)) {
+		if (!rz_core_types_typedef_info(core, typename)) {
 			eprintf("Can't find typedef");
 			return RZ_CMD_STATUS_ERROR;
 		}
@@ -2232,11 +1343,12 @@ RZ_IPI RzCmdStatus rz_type_list_typedef_handler(RzCore *core, int argc, const ch
 
 RZ_IPI RzCmdStatus rz_type_typedef_c_handler(RzCore *core, int argc, const char **argv) {
 	const char *typename = argc > 1 ? argv[1] : NULL;
+	Sdb *TDB = core->analysis->sdb_types;
 	if (!typename) {
-		rz_core_list_typename_alias_c(core, NULL);
+		rz_types_typedef_print_c(TDB, NULL);
 		return RZ_CMD_STATUS_OK;
 	}
-	rz_core_list_typename_alias_c(core, typename);
+	rz_types_typedef_print_c(TDB, typename);
 	return RZ_CMD_STATUS_OK;
 }
 
@@ -2244,14 +1356,14 @@ RZ_IPI RzCmdStatus rz_type_list_union_handler(RzCore *core, int argc, const char
 	const char *typename = argc > 1 ? argv[1] : NULL;
 	Sdb *TDB = core->analysis->sdb_types;
 	if (typename) {
-		type_show_format(core, typename, mode);
+		rz_core_types_show_format(core, typename, mode);
 	} else {
 		if (mode == RZ_OUTPUT_MODE_RIZIN) {
-			print_all_union_format(core, TDB);
+			rz_core_types_union_print_format_all(core, TDB);
 		} else if (mode == RZ_OUTPUT_MODE_JSON) {
-			print_struct_union_list_json(TDB, stdifunion);
+			rz_types_union_print_json(TDB);
 		} else {
-			print_keys(TDB, core, stdifunion, printkey_cb, false);
+			rz_types_union_print_sdb(TDB);
 		}
 	}
 	return RZ_CMD_STATUS_OK;
@@ -2260,13 +1372,13 @@ RZ_IPI RzCmdStatus rz_type_list_union_handler(RzCore *core, int argc, const char
 RZ_IPI RzCmdStatus rz_type_union_c_handler(RzCore *core, int argc, const char **argv) {
 	const char *typename = argc > 1 ? argv[1] : NULL;
 	Sdb *TDB = core->analysis->sdb_types;
-	print_struct_union_in_c_format(TDB, stdifunion, typename, true);
+	rz_types_union_print_c(TDB, typename, true);
 	return RZ_CMD_STATUS_OK;
 }
 
 RZ_IPI RzCmdStatus rz_type_union_c_nl_handler(RzCore *core, int argc, const char **argv) {
 	const char *typename = argc > 1 ? argv[1] : NULL;
 	Sdb *TDB = core->analysis->sdb_types;
-	print_struct_union_in_c_format(TDB, stdifunion, typename, false);
+	rz_types_union_print_c(TDB, typename, false);
 	return RZ_CMD_STATUS_OK;
 }
