+++ librz/core/core_private.h
@@ -7,6 +7,9 @@
 RZ_IPI void rz_core_kuery_print(RzCore *core, const char *k);
 RZ_IPI int rz_output_mode_to_char(RzOutputMode mode);
 
+RZ_IPI int bb_cmpaddr(const void *_a, const void *_b);
+RZ_IPI int fcn_cmpaddr(const void *_a, const void *_b);
+
 RZ_IPI int rz_core_analysis_set_reg(RzCore *core, const char *regname, ut64 val);
 RZ_IPI void rz_core_analysis_esil_init(RzCore *core);
 RZ_IPI void rz_core_analysis_esil_reinit(RzCore *core);
@@ -23,8 +26,6 @@ RZ_IPI void rz_core_analysis_esil_emulate_bb(RzCore *core);
 RZ_IPI void rz_core_analysis_esil_default(RzCore *core);
 
 RZ_IPI bool rz_core_analysis_var_rename(RzCore *core, const char *name, const char *newname);
-RZ_IPI RzList *rz_core_analysis_calling_conventions(RzCore *core);
-RZ_IPI void rz_core_analysis_calling_conventions_print(RzCore *core);
 RZ_IPI char *rz_core_analysis_function_signature(RzCore *core, RzOutputMode mode, char *fcn_name);
 RZ_IPI bool rz_core_analysis_everything(RzCore *core, bool experimental, char *dh_orig);
 RZ_IPI bool rz_core_analysis_function_delete_var(RzCore *core, RzAnalysisFunction *fcn, RzAnalysisVarKind kind, const char *id);
@@ -39,9 +40,44 @@ RZ_IPI void rz_core_analysis_bbs_info_print(RzCore *core, RzAnalysisFunction *fc
 RZ_IPI void rz_core_analysis_bb_info_print(RzCore *core, RzAnalysisBlock *bb, ut64 addr, RzOutputMode mode);
 RZ_IPI void rz_core_analysis_function_until(RzCore *core, ut64 addr_end);
 RZ_IPI void rz_core_analysis_value_pointers(RzCore *core, RzOutputMode mode);
-RZ_IPI RzList *rz_core_analysis_noreturn(RzCore *core);
-RZ_IPI void rz_core_analysis_noreturn_print(RzCore *core, RzOutputMode mode);
 
+/* ctypes.c */
+RZ_IPI RzList *rz_types_calling_conventions(Sdb *db);
+RZ_IPI void rz_core_types_calling_conventions_print(RzCore *core, RzOutputMode mode);
+RZ_IPI RzList *rz_types_enums(Sdb *db);
+RZ_IPI void rz_core_types_enum_print(RzCore *core, const char *enum_name, RzOutputMode mode, PJ *pj);
+RZ_IPI void rz_core_types_enum_print_all(RzCore *core, RzOutputMode mode);
+RZ_IPI void rz_types_enum_print_c(Sdb *TDB, const char *arg, bool multiline);
+RZ_IPI bool rz_core_types_typedef_info(RzCore *core, const char *name);
+RZ_IPI void rz_types_typedef_print_c(Sdb *TDB, const char *typedef_name);
+RZ_IPI void rz_core_list_loaded_typedefs(RzCore *core, RzOutputMode mode);
+
+// Structured types JSON
+RZ_IPI void rz_types_structured_print_json(Sdb *TDB, SdbList *l);
+RZ_IPI void rz_types_union_print_json(Sdb *TDB);
+RZ_IPI void rz_types_struct_print_json(Sdb *TDB);
+// Structured types SDB
+RZ_IPI void rz_types_structured_print_sdb(Sdb *TDB, SdbList *l);
+RZ_IPI void rz_types_union_print_sdb(Sdb *TDB);
+RZ_IPI void rz_types_struct_print_sdb(Sdb *TDB);
+// Structured types C format
+RZ_IPI void rz_types_union_print_c(Sdb *TDB, const char *name, bool multiline);
+RZ_IPI void rz_types_struct_print_c(Sdb *TDB, const char *name, bool multiline);
+RZ_IPI void rz_types_function_print(Sdb *TDB, const char *function, RzOutputMode mode, PJ *pj);
+RZ_IPI void rz_core_types_function_print_all(RzCore *core, RzOutputMode mode);
+RZ_IPI RzList *rz_types_function_noreturn(Sdb *db);
+RZ_IPI void rz_core_types_function_noreturn_print(RzCore *core, RzOutputMode mode);
+RZ_IPI void rz_core_types_show_format(RzCore *core, const char *name, RzOutputMode mode);
+RZ_IPI void rz_core_types_struct_print_format_all(RzCore *core, Sdb *TDB);
+RZ_IPI void rz_core_types_union_print_format_all(RzCore *core, Sdb *TDB);
+RZ_IPI RzList *rz_types_links(Sdb *db);
+RZ_IPI void rz_core_types_link_print(RzCore *core, const char *type, ut64 addr, RzOutputMode mode);
+RZ_IPI void rz_core_types_link_print_all(RzCore *core, RzOutputMode mode);
+RZ_IPI void rz_core_types_link(RzCore *core, const char *type, ut64 addr);
+RZ_IPI void rz_core_types_link_show(RzCore *core, ut64 addr);
+RZ_IPI void rz_core_types_print_all(RzCore *core, RzOutputMode mode);
+
+/* agraph.c */
 RZ_IPI void rz_core_agraph_add_node(RzCore *core, const char *title, const char *body, int color);
 RZ_IPI void rz_core_agraph_del_node(RzCore *core, const char *title);
 RZ_IPI void rz_core_agraph_add_edge(RzCore *core, const char *un, const char *vn);
