+++ test/db/cmd/cmd_ara
@@ -90,17 +90,17 @@ tcc `arcc`
 afcl
 EOF
 EXPECT=<<EOF
-swift
 amd64
 amd64syscall
 ms
 reg
+swift
 r0 reg(r0, r1, r2, r3)
 rdi reg(rdi, rsi, rdx, rcx)
-swift
 amd64
 amd64syscall
 ms
 reg
+swift
 EOF
 RUN
